pyproject.toml
[tool.poetry]
name = "postgresql-db-generator"
version = "0.1.0"
description = "Generate multiple PostgreSQL databases and users with restricted privileges"
authors = ["Henri Aïdasso <ahenrij@gmail.com>"]
license = "LICENCE"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
